finish rebase
