+++ examples/dynamic_api.rs
@@ -2,10 +2,10 @@
 //! to `kubectl get all --all-namespaces`.
 
 use kube::{
-    api::{Api, ApiResource, DynamicObject, Resource, ResourceExt},
-    client::{Client, Discovery},
+    api::{Api, DynamicObject, ResourceExt},
+    client::{Client, Discovery, Scope},
 };
-use log::{info, warn};
+use log::info;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
@@ -22,30 +22,23 @@ async fn main() -> anyhow::Result<()> {
 
     for group in discovery.groups() {
         let ver = group.preferred_version_or_guess();
-        for gvk in group.resources_by_version(ver) {
-            let kind = DynamicObject::kind(&gvk);
-            let (_, raw_resource) = discovery
-                .resolve_group_version_kind(group.name(), ver, &kind)
-                .unwrap();
-            let api: Api<DynamicObject> = if raw_resource.namespaced {
+        for (api_res, extras) in group.resources_by_version(ver) {
+            if !extras.operations.list {
+                continue;
+            }
+            let api: Api<DynamicObject> = if extras.scope == Scope::Namespaced {
                 if let Some(ns) = &ns_filter {
-                    Api::namespaced_with(client.clone(), ns, &gvk)
+                    Api::namespaced_with(client.clone(), ns, &api_res)
                 } else {
-                    Api::all_with(client.clone(), &gvk)
+                    Api::all_with(client.clone(), &api_res)
                 }
             } else {
-                Api::all_with(client.clone(), &gvk)
+                Api::all_with(client.clone(), &api_res)
             };
 
-            info!("{}/{} : {}", group.name(), ver, kind);
+            info!("{}/{} : {}", group.name(), ver, api_res.kind);
 
-            let list = match api.list(&Default::default()).await {
-                Ok(l) => l,
-                Err(e) => {
-                    warn!("Failed to list: {:#}", e);
-                    continue;
-                }
-            };
+            let list = api.list(&Default::default()).await?;
             for item in list.items {
                 let name = item.name();
                 let ns = item.metadata.namespace.map(|s| s + "/").unwrap_or_default();
